\name{paquetevariado-package}
\alias{paquetevariado-package}
\alias{paquetevariado}
\docType{package}
\title{
\packageTitle{paquetevariado}
}
\description{
\packageDescription{paquetevariado}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{paquetevariado}
\packageIndices{paquetevariado}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{paquetevariado}

Maintainer: \packageMaintainer{paquetevariado}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
